requirements.txt
alembic==1.6.5
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bullet==2.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
classproperties==0.2.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
colorama==0.4.4
configparser==5.0.2
crashtest==0.3.1
crayons==0.4.0
datafiles==0.15
decorator==5.0.9
fake-useragent==0.1.11
Faker==4.18.0
Flask==1.1.2
Flask-API==3.0.post1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gitman==2.3.2
greenlet==1.1.0
idna==2.10
inflection==0.4.0
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
jellyfish==0.6.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
minilog==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
parse==1.19.0
parso==0.8.2
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pomace==0.8.7
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylev==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
splinter==0.14.0
SQLAlchemy==1.4.18
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
us==2.0.2
wcwidth==0.2.5
webdriver-manager==2.5.3
Werkzeug==2.0.1
zipcodes==1.1.2
